+++ src/app/(pages)/about/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function page() {
+  return (
+    <div>About Page</div>
+  )
+}
+
+export default page
