+++ src/app/Components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/utils/utils"
 
 const Card = React.forwardRef<
   HTMLDivElement,
