+++ src/app/layout.tsx
@@ -1,26 +1,14 @@
 import type { Metadata } from "next";
-import localFont from "next/font/local";
 import "./globals.css";
 import "@xyflow/react/dist/style.css";
-import Navbar from "./Components/navbar";
+import Navbar from "./components/Navbar";
+import StoreProvider from "./lib/store/StoreProvider";
 
-import WorkFLowSideBar from "./Components/WorkFLowSideBar";
-import StoreProvider from "./StoreProvider";
 
-const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
-  variable: "--font-geist-sans",
-  weight: "100 900",
-});
-const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
-  variable: "--font-geist-mono",
-  weight: "100 900",
-});
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Asset Management Tool",
+  description: "By Lyktek.io",
 };
 
 export default function RootLayout({
@@ -31,7 +19,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased w-full h-screen overflow-hidden`}
+        className={`antialiased w-full h-screen overflow-hidden`}
       >
         <StoreProvider>
           <Navbar />
